Add some extra info to ddns-service
To check if the DDNS service is working properly I manually edited a DNS
record after starting this service to see if it would update back again
on the next interval. I think this was how it worked with the previous
DDNS service I used.

Reading some more upon the code I now understand this service does not
compare the current DNS record value with the current public IP address,
but rather store it in memory and only updates the DNS entry when it
does not match the current public IP address.

I noticed in the closed issues section another user with probably the
same expectations. Hopefully this commit helps users understand the
functionality better.